Generate Data Matrix (X) From a Linear Regression Model.

Generates random data matrix from a \(k\)-variable linear regression model.

gendat_linreg_X(n, k, constant = TRUE, rFUN_X = rnorm, ...)

Arguments

n
Sample size.

k
Number of regressors.

constant
Logical. An option to include or to exclude the vector of constants. If TRUE, the vector of constants is included. If FALSE, the vector of constants is excluded.

rFUN_X
The distribution function used to generate values of \(\mathbf{X}\). The default value is rnorm for the Gaussian probability density function.

...
Arguments to pass to rFUN_X.

Value

If constant = TRUE, returns an \(n \times k\) numeric matrix where the first column consists of 1s. If constant = FALSE, returns an \(n \times k - 1\) numeric matrix.

Details

Randomly generates the data matrix (\(\mathbf{X}\)), that is an \(n \times k\) dimensional matrix of \(n\) observations of \(k\) regressors, which includes a regressor whose value is 1 for each observation. The data generating function is supplied by the argument rFUN_X (rnorm is the default value). Additional arguments to rFUN_X are supplied using the ... argument. The data matix is also called the design matrix and model matrix.

See also

Examples

X <- gendat_linreg_X(
  n = 100,
  k = 3,
  constant = TRUE,
  rFUN_X = rnorm,
  mean = 100,
  sd = 15
)