Generate Regressand Data (y) From a Linear Regression Model.

Generates regressand data from a \(k\)-variable linear regression model.

gendat_linreg_y(X, beta, rFUN_y = rnorm, ...)

Arguments

X
The data matrix, that is an \(n \times k\) matrix of \(n\) observations of \(k\) regressors, which includes a regressor whose value is 1 for each observation. Also called the design matrix and model matrix.

beta
\(k \times 1\) vector of \(k\) regression parameters.

rFUN_y
The distribution function used to generate values of the residuals \(\epsilon\). The default value is rnorm for the Gaussian probability density function.

...
Arguments to pass to rFUN_y.

Details

Randomly generates the regressand data (\(\mathbf{y}\)) using specified population parameters defined by \(\mathbf{y}_{n \times 1} = \mathbf{X}_{n \times k} \boldsymbol{\beta}_{k \times 1} + \boldsymbol{\epsilon}_{n \times 1} \). The distribution of \(\epsilon\) is supplied by the argument rFUN_y (rnorm is the default value). Additional arguments to rFUN_y are supplied using the ... argument. By default, \(\epsilon\) is assumed to be normally distributed with a mean of 0 and a variance of 1 (\( \mathcal{N} \sim \left( \mu_{\epsilon} = 0, \sigma_{\epsilon}^{2} = 1 \right) \)).

See also

Examples

X <- gendat_linreg_X(
  n = 100,
  k = 3,
  constant = TRUE,
  rFUN_X = rnorm,
  mean = 0,
  sd = 1
)
y <- gendat_linreg_y(
  X = X,
  beta = c(.5, .5, .5),
  rFUN_y = rnorm,
  mean = 0,
  sd = 1
)