LISREL Structural Equations with Latent Variables.

Model-implied variance-covariance matrix (\(\boldsymbol{\Sigma} \left( \boldsymbol{\theta} \right)\)) using the LISREL notation for structural equations with latent variables.

lisrel(LY, LX, TE, TD, BE, I, GA, PS, PH)

Arguments

LY
\(\boldsymbol{\Lambda}_{\mathbf{y}}\) \(p \times m\) matrix of factor loadings (\(\boldsymbol{\lambda}\)). \(p\) is the number of observed indicators (\(\mathbf{y}\)) and \(m\) is the number of latent endogenous variables (\(\boldsymbol{\eta}\)).

LX
\(\boldsymbol{\Lambda}_{\mathbf{x}}\) \(q \times n\) matrix of factor loadings (\(\boldsymbol{\lambda}\)). \(q\) is the number of observed indicators (\(\mathbf{x}\)) and \(n\) is the number of latent exogenous variables (\(\boldsymbol{\xi}\)).

TE
\(\boldsymbol{\Theta_{\boldsymbol{\epsilon}}}\) \(p \times p\) matrix of residual variances and covariances for \(\mathbf{y}\) (\(\boldsymbol{\epsilon}\)).

TD
\(\boldsymbol{\Theta_{\boldsymbol{\delta}}}\) \(q \times q\) matrix of residual variances and covariances for \(\mathbf{x}\) (\(\boldsymbol{\delta}\)).

BE
\(\mathbf{B}_{m \times m}\) coefficient matrix for endogenous variables.

I
\(\mathbf{I}_{m \times m}\) identity matrix.

GA
\(\boldsymbol{\Gamma}_{m \times n}\) coefficient matrix for exogenous variables.

PS
\(\boldsymbol{\Psi}_{m \times m}\) variance-covariance of \(\boldsymbol{\zeta}\). \(\boldsymbol{\zeta}\) is a matrix of residual variances and covariances in regression equations.

PH
\(\boldsymbol{\Phi}_{n \times n}\) variance-covariance matrix of \(\boldsymbol{\xi}\).

Value

Returns the model-implied variance-covariance matrix (\(\boldsymbol{\Sigma} \left( \boldsymbol{\theta} \right)\)) derived from the \(\boldsymbol{\Lambda}_{\mathbf{y}}\) (LY), \(\boldsymbol{\Lambda}_{\mathbf{x}}\) (LX), \(\boldsymbol{\Theta}_{\boldsymbol{\epsilon}}\) (TE), \(\boldsymbol{\Theta}_{\boldsymbol{\delta}}\) (TD), \(\mathbf{B}\) (BE), \(\mathbf{I}\) (I), \(\boldsymbol{\Gamma}\) (GA), \(\boldsymbol{\Psi}\) (PS), and \(\boldsymbol{\Phi}\) (PH) matrices.

Details

Combines \(\boldsymbol{\Sigma}_{\mathbf{yy}} \left( \boldsymbol{\theta} \right)\) (from lisrel_yy), \(\boldsymbol{\Sigma}_{\mathbf{yx}} \left( \boldsymbol{\theta} \right)\) (from lisrel_yx), \(\boldsymbol{\Sigma}_{\mathbf{xy}} \left( \boldsymbol{\theta} \right)\) (from lisrel_xy), and \(\boldsymbol{\Sigma}_{\mathbf{xx}} \left( \boldsymbol{\theta} \right)\) (from lisrel_xx) to produce \(\boldsymbol{\Sigma} \left( \boldsymbol{\theta} \right)\) . The variance-covariance matrices are derived using the following equations $$\boldsymbol{\Sigma}_{\mathbf{yy}} \left( \boldsymbol{\theta} \right) = \boldsymbol{\Lambda}_{\mathbf{y}} \left( \mathbf{I} - \mathbf{B} \right)^{-1} \left( \boldsymbol{\Gamma} \boldsymbol{\Phi} \boldsymbol{\Gamma}^{T} + \boldsymbol{\Psi} \right) \left[ \left( \mathbf{I} - \mathbf{B} \right)^{-1} \right]^{T} \boldsymbol{\Lambda}_{\mathbf{y}}^{T} + \boldsymbol{\Theta}_{\boldsymbol{\epsilon}} $$ $$\boldsymbol{\Sigma}_{\mathbf{yx}} \left( \boldsymbol{\theta} \right) = \boldsymbol{\Lambda}_{\mathbf{y}} \left( \mathbf{I} - \mathbf{B} \right)^{-1} \boldsymbol{\Gamma} \boldsymbol{\Phi} \boldsymbol{\Lambda}_{\mathbf{x}}^{T} $$ $$\boldsymbol{\Sigma}_{\mathbf{xy}} \left( \boldsymbol{\theta} \right) = \boldsymbol{\Lambda}_{\mathbf{x}} \boldsymbol{\Phi} \boldsymbol{\Gamma}^{T} \left[ \left( \mathbf{I} - \mathbf{B} \right)^{-1} \right]^{T} \boldsymbol{\Lambda}_{\mathbf{y}}^{T} $$ $$\boldsymbol{\Sigma}_{\mathbf{xx}} \left( \boldsymbol{\theta} \right) = \boldsymbol{\Lambda}_{\mathbf{x}} \boldsymbol{\Phi} \boldsymbol{\Lambda}_{\mathbf{y}}^{T} + \boldsymbol{\Theta}_{\boldsymbol{\delta}} . $$

References

Bollen, K. A. (1989). Structural equations with latent variables. New York: Wiley.

Jöreskog, K. G., & Sörbom, D. (1996). Lisrel 8: User's reference guide (2nd ed.). Scientific Software.

See also