Simple Mediation Model

Estimates the indirect effect in a simple mediation model, that is the product of \(\alpha\) and \(\beta\) from \(M_i = \delta_M + \alpha X_i + \epsilon_{M_i}\) and \(Y_i = \delta_Y + \tau^{\prime} X_i + \beta M_i + \epsilon_{Y_i}\).

med_simple(data, scale = FALSE, minimal = TRUE, s2 = FALSE, s2_est = "both")

Arguments

data
A matrix with variables X, M, and Y.

scale
Logical. If TRUE, scales the data before fitting the model.

minimal
Logical. If TRUE, returns the indirect effect of X on Y through M. If FALSE, returns all the regression coefficients estimated.

s2
Logical. If TRUE, estimates residual variance.

s2_est
String. Residual variance estimator. If "both", returns both OLS and ML estimates as a vector. If "ols", returns OLS estimate. If "ml", returns ML estimate. Ignored if s2 = FALSE.

Examples

Sigma <- matrix(
  data = c(
    225, 112.50, 56.25,
    112.5, 225, 112.5,
    56.25, 112.50, 225
  ),
  ncol = 3
)
mu <- c(100, 100, 100)
data <- gendat_mvn(
  n = 100,
  Sigma = Sigma,
  mu = mu
)
med_simple(data = data, minimal = FALSE, s2 = TRUE)

#>          d_M            a          d_Y      c_prime            b   s2_e_M_ols 
#>  60.00241119   0.40167417  44.78207983   0.05739135   0.49585264 196.39520420 
#>    s2_e_M_ml   s2_e_Y_ols    s2_e_Y_ml 
#> 192.46730011 177.03060055 171.71968253